add level.open/close.blind roles
After the discussion in the boards it seems there is no consensus on how level.blind has to be interpreted. This makes it hard for visualisations and other adapters that try to make things more smart to control them or interpret readings from value.blind. So the (middle ground)[https://forum.iobroker.net/topic/33995/umfrage-rolladenposition-in-was-ist-logischer/85] was to add two roles which make the interpretation more clear. 

I added those new roles here for level.open/close.blind, level.open/close.tilt, the same for value and some buttons. I plan on using them in a future PR for type-detector.